+++ ethcore/src/engines/null_engine.rs
@@ -22,7 +22,7 @@ use spec::CommonParams;
 use evm::Schedule;
 use env_info::EnvInfo;
 
-/// An engine which does not provide any consensus mechanism.
+/// An engine which does not provide any consensus mechanism and does not seal blocks.
 pub struct NullEngine {
 	params: CommonParams,
 	builtins: BTreeMap<Address, Builtin>,
